[Resolves #1] Create urls table and model
# Feature Title

Create urls table and model

# Changes made

- Created migration that will allow storing shortened urls and respective index
- Created operation to create a shortened url, using trailblazer
- Created small library to transform an integer into a string base on the 26 alphabet characters
- Prepared trailblazer configuration

# How does the solution address the issue

The operation allows us to have a self calling class that will create a new record on the newly created table, provided that the contract is valid.
The uniqueness of the URL is guaranteed by the uniqueness of the id of the record.
Converting it into a string is merely following nowadays URL shortening services. The algorithm can later be changed if deemed necessary.