[Resolves #5] Collect basic url statistics
# Feature Title

Collect basic url statistics

# Changes made

- Added storext gem which adds features on top of ActiveRecord::Store.store_accessor
- Added step to track statistics on visit operation
- Added migration to create url_statistics table
- Adding UUID to the user's cookies if not set yet

# How does the solution address the issue

Adding the table url statistics and associate it with the url table allows us to keep track of the statistics of a specific url.
An entry is recorded everytime a user tries to visit a specific URL, in case it is found in the database.
Using ActiveRecord::Store allows us to easily expand the tracked data, which currently includes:
  - user agent
  - host
  - accept language
  - remote address

The user is identified via cookie (the uuid). This can later be used to increase the data association with google analytics by sending the same uuid to google analytics.